+++ R/cds_delete.R
@@ -1,3 +1,4 @@
+
 #' CDS delete request
 #'
 #' Deletes a staged download from the queue
@@ -5,7 +6,8 @@
 #' @param user character, user ID to sign up for the CDS data service and
 #' used to retrieve the token set by \code{\link[ecmwfr]{cds_set_key}}.
 #' Can also be \code{NULL} if you preferr to use the \code{.cdsapirc} file.
-#' @param url task url for request
+#' @param url task url for request OR the character with the request ID.
+#' If an ID is given the url will be auto-generated.
 #' @param verbose show feedback on processing
 #' @keywords data download, climate, re-analysis
 #' @seealso \code{\link[ecmwfr]{cds_set_key}}
@@ -17,10 +19,11 @@ cds_delete <- function(user, url, verbose = TRUE){
 
   # check the login credentials
   if(missing(user) | missing(url)){
-    stop("Please provide ECMWF login email / url!")
+    stop("Please provide ECMWF login email and url/id (see manual page)!")
   }
 
-  # get key
+  # get key. If 'user == NULL' load user login information from
+  # '~/.cdsapirc' file. Else load key via local keyring.
   if(is.null(user)) {
     tmp <- cds_key_from_file(verbose = verbose)
     user <- tmp$user; key <- tmp$key; rm(tmp)